Move persistence.js to app/session/ dir
